fixed edit and delete events

+++ app/controllers/api/events_controller.rb
@@ -35,6 +35,14 @@ class Api::EventsController < ApplicationController
           )
       event_tag.save
     end
+    images_input = params[:image_urls].split(", ")
+    images_input.each do |image_url|
+      image = Image.new(
+        url: image_url,
+        event_id: @event.id
+        )
+      image.save
+    end
     render "show.json.jbuilder"
   end
 
@@ -45,38 +53,60 @@ class Api::EventsController < ApplicationController
 
   def update
     @event = Event.find_by_id(params[:id])
-    @event.title = params[:title] || @event.title
-    @event.address = params[:address] || @event.address
-    @event.description = params[:description] || @event.desription
-    @event.date = params[:date] || @event.date
-    if params[:tag_names]
-      tags_input = params[:tag_names].split(", ")
-      tags_input.each do |tag_name|
-        tag = Tag.find_by(name: tag_name)
-        if !tag
-          tag = Tag.new(
-            name: tag_name
-            )
-          tag.save
+    if @event.creator.id == current_user.id
+      @event.title = params[:title] || @event.title
+      @event.address = params[:address] || @event.address
+      @event.description = params[:description] || @event.description
+      @event.date = params[:date] || @event.date
+      @event.save
+      if params[:tag_names]
+        tags_input = params[:tag_names].split(", ")
+        tags_input.each do |tag_name|
+          tag = Tag.find_by(name: tag_name)
+          if !tag
+            tag = Tag.new(
+              name: tag_name
+              )
+            tag.save
+          end
+          event_tag = EventTag.new(
+              event_id: @event[:id],
+              tag_id: tag[:id]
+              )
+          event_tag.save
         end
-        event_tag = EventTag.new(
-          event_id: @event[:id],
-          tag_id: tag[:id]
-          )
-        event_tag.save
       end
+      if params[:image_urls]
+        images_input = params[:image_urls].split(", ")
+        images_input.each do |image_url|
+          image = Image.new(
+            url: image_url,
+            event_id: @event.id
+            )
+          image.save
+        end
+      end
+      render "show.json.jbuilder"
+    else
+      render json: {}, status: :unauthorized
     end
-    render "show.json.jbuilder"
   end
 
   def destroy
     event = Event.find_by_id(params[:id])
-    event.event_tags.each do |event_tag|
-      event_tag.destroy
+    if event.creator.id == current_user.id
+      event.event_tags.each do |event_tag|
+        event_tag.destroy
+      end
+      event.user_events.each do |user_event|
+        user_event.destroy
+      end
+      event.images.each do |image|
+        image.destroy
+      end
+      event.destroy
+    else 
+      render json: {}, status: :unauthorized
     end
-    event.user_events.each do |user_event|
-      user_event.destroy
-    end
-    event.destroy
   end
 end
